fix: separate user atom and user hooks and placed user atom inside state module
