+++ packages/framework/src/state/index.ts
@@ -2,3 +2,4 @@ export * from "./widget";
 export * from "./screen";
 export * from "./application";
 export * from "./platform";
+export * from "./user";
